+++ src/store_grid_data.c
@@ -39,21 +39,21 @@ void	store_max_alt(t_grid_data *grid_data, t_pixel **grid)
 
 void	store_spacing(t_grid_data *grid_data, t_pixel **grid)
 {
-	float	height;
-	float	width;
+	double	height;
+	double	width;
 
 	grid_data->x_max = (grid_data->width - 1) * 0.71;
 	grid_data->x_min = (0 - (grid_data->height - 1)) * 0.71;
 	grid_data->y_max = -0.82 * (
-			(float)grid[grid_data->height - 1][grid_data->width - 1].z / (
+			(double)grid[grid_data->height - 1][grid_data->width - 1].z / (
 				(grid_data->alt_max - grid_data->alt_min)
-				/ ((float)(grid_data->height + grid_data->width) / 16)))
+				/ ((double)(grid_data->height + grid_data->width) / 16)))
 		+ 0.41 * (grid_data->width - 1 + (grid_data->height - 1));
-	grid_data->y_min = -0.82 * ((float)grid[0][0].z / (
+	grid_data->y_min = -0.82 * ((double)grid[0][0].z / (
 				(grid_data->alt_max - grid_data->alt_min)
-				/ ((float)(grid_data->height + grid_data->width) / 16)));
-	height = ((float)HEIGHT * ZOOM) / (grid_data->y_max - grid_data->y_min);
-	width = ((float)WIDTH * ZOOM) / (grid_data->x_max - grid_data->x_min);
+				/ ((double)(grid_data->height + grid_data->width) / 16)));
+	height = ((double)HEIGHT * ZOOM) / (grid_data->y_max - grid_data->y_min);
+	width = ((double)WIDTH * ZOOM) / (grid_data->x_max - grid_data->x_min);
 	if (height < width)
 		grid_data->spacing = height;
 	else
@@ -63,20 +63,20 @@ void	store_spacing(t_grid_data *grid_data, t_pixel **grid)
 static void	get_color(t_grid_data *grid_data, t_pixel *cell)
 {
 	unsigned int	mask;
-	float			ratio;
+	double			ratio;
 	int				color_ext;
 	int				tmp_neut;
 	int				tmp_ext;
 
 	color_ext = COLOR_H;
 	if (cell->z > 0)
-		ratio = (float)cell->z / grid_data->alt_max;
+		ratio = (double)cell->z / grid_data->alt_max;
 	else if (cell->z == 0)
 		ratio = 0;
 	else if (cell->z < 0)
 	{
 		color_ext = COLOR_L;
-		ratio = (float)cell->z / grid_data->alt_min;
+		ratio = (double)cell->z / grid_data->alt_min;
 	}
 	mask = BLUE;
 	cell->color = 0;
@@ -102,15 +102,15 @@ void	store_iso_pos(t_grid_data *grid_data, t_pixel **grid)
 		{
 			if (grid[y][x].color == 0)
 				get_color(grid_data, &grid[y][x]);
-			grid[y][x].x = (float)((x - y) * 0.71 - grid_data->x_min)
-				*grid_data->spacing + (WIDTH - (float)WIDTH * ZOOM) / 2;
-			grid[y][x].y = (-0.82 * ((float)grid[y][x].z / (
+			grid[y][x].x = (double)((x - y) * 0.71 - grid_data->x_min)
+				*grid_data->spacing + (WIDTH - (double)WIDTH * ZOOM) / 2;
+			grid[y][x].y = (-0.82 * ((double)grid[y][x].z / (
 							(grid_data->alt_max - grid_data->alt_min)
-							/ ((float)(grid_data->height + grid_data->width)
+							/ ((double)(grid_data->height + grid_data->width)
 								/ 16)))
 					+ 0.41 * (x + y) - grid_data->y_min)
 				* grid_data->spacing
-				+ (HEIGHT - (float)HEIGHT * ZOOM) / 2;
+				+ (HEIGHT - (double)HEIGHT * ZOOM) / 2;
 			x++;
 		}
 		y++;
